PlotSimultaneous projection speedup
Implemented a fix for plotting simultaneous PDFs when using ProjWData command arguments. In summed plots this would lead to very slow projection, even if requesting binned projection. This version automatically detects a request for binned projection. If so, it generates a binned dataset that will be used for projection, forcing fast binned projection. RooSuperCategories are detected and handled automatically. Allocated memory will be deleted afterwards.